Static cast on bool to int conversion
Fixes clang-tidy build failure introduced by d837bf554c

+++ src/gromacs/mdlib/mdgraph_gpu_impl.cu
@@ -171,7 +171,7 @@ void MdGpuGraph::Impl::disableForDomainIfAnyPpRankHasCpuForces(bool disableGraph
     if (havePPDomainDecomposition_)
     {
         // If disabled on any domain, disable on all domains
-        int disableGraphAcrossAllPpRanksReductionInput  = disableGraphAcrossAllPpRanks;
+        int disableGraphAcrossAllPpRanksReductionInput = static_cast<int>(disableGraphAcrossAllPpRanks);
         int disableGraphAcrossAllPpRanksReductionOutput = 0;
         MPI_Allreduce(&disableGraphAcrossAllPpRanksReductionInput,
                       &disableGraphAcrossAllPpRanksReductionOutput,
